+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
-**3.0.7 - 09/23/24**
+**3.0.7 - 09/24/24**
 
-  - Enable population manager and population view methods to accept strings
+  - Fix mypy errors: vivarium/framework/lifecycle.py
   - Fix mypy errors: vivarium/framework/population/manager.py
   - Fix mypy errors: vivarium/framework/population/population_view.py
+  - Enable population manager and population view methods to accept strings
 
 **3.0.6 - 09/20/24**